fixed migration

+++ db/migrate/20240320_add_image_url_to_lodgings.rb
@@ -1,6 +1,11 @@
 class AddImageUrlToLodgings < ActiveRecord::Migration[7.1]
   def change
-    add_column :lodgings, :image_url, :string
-    remove_column :lodgings, :photo_reference, :string
+    unless column_exists?(:lodgings, :image_url)
+      add_column :lodgings, :image_url, :string
+    end
+    
+    if column_exists?(:lodgings, :photo_reference)
+      remove_column :lodgings, :photo_reference, :string
+    end
   end
 end
